requirements.txt
moviepy >= 1.0.3
Pillow >= 10.1.0
numpy >= 1.26.2
configobj >= 5.0.8
openai >= 0.28.1
google-cloud-texttospeech >= 2.15.0